+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Chores
 
+- Fixed Language components in Storybook to have it's own mocked store ([PR 3063](https://github.com/input-output-hk/daedalus/pull/3063))
 - Updated `cardano-node` configs for Preprod respin ([PR 3070](https://github.com/input-output-hk/daedalus/pull/3070))
 - Fixed the properties and providers for the Send stories ([PR 3047](https://github.com/input-output-hk/daedalus/pull/3047))
 - Fixed locale in Storybook ([PR 3062](https://github.com/input-output-hk/daedalus/pull/3062))